changed pgx version to pgx4 batch insert with RETURNING works normally created method for create posts in thread

+++ configs/database/sql/database.sql
@@ -89,7 +89,7 @@ CREATE UNLOGGED TABLE IF NOT EXISTS posts
     forum_slug      CITEXT                                                NOT NULL,
     is_edited       BOOLEAN                     DEFAULT FALSE             NOT NULL,
     message         TEXT                                                  NOT NULL,
-    parent          INTEGER,
+    parent          BIGINT,
     created         TIMESTAMP(3) WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP NOT NULL,
     path            BIGINT[]                                              NOT NULL,
